Task13/authorization (#11)
* add credentials

* Add authorization by Cancan

* Cancan authorization in views

* Small fix

* Fix links

* Fixes

+++ app/controllers/links_controller.rb
@@ -2,9 +2,7 @@ class LinksController < ApplicationController
   before_action :authenticate_user!, only: :destroy
 
   def destroy
-    if current_user&.author_of?(link.linkable)
-      link.destroy
-    end
+    link.destroy if authorize! :destroy, link.linkable
   end
 
   private
